Export schema.
