+++ package.json
@@ -4,7 +4,7 @@
   "description": "Server component of the paper-walrus project.",
   "main": "src/index.js",
   "scripts": {
-    "start": "node ."
+    "start": "node serve.js"
   },
   "repository": {
     "type": "git",
